+++ content/pages/news.md
@@ -2,3 +2,11 @@ Title: Lucene™ Project News
 URL: news.html
 save_as: news.html
 template: lucene/tlp/news
+
+You may also read Lucene news as ATOM feeds:
+
+* [TLP news feed](/feeds/news.atom.xml).
+* [All news feed](/feeds/all.atom.xml).
+* [Lucene core news feed](/feeds/core/news.atom.xml).
+* [Solr regular news feed](/feeds/solr/news.atom.xml).
+* [Solr security news feed](/feeds/solr/security.atom.xml).