+++ content/pages/solr/news.md
@@ -2,3 +2,5 @@ Title: Solr News
 URL: solr/news.html
 save_as: solr/news.html
 template: solr/news
+
+You may also read these news as an [ATOM feed](/feeds/solr/news.atom.xml).